Move liquibase update functions to separate class.
