+++ libs/db/db-admin-impl/src/main/kotlin/net/corda/db/admin/impl/LiquibaseSchemaMigratorImpl.kt
@@ -1,26 +1,14 @@
 package net.corda.db.admin.impl
 
 import liquibase.Contexts
-import liquibase.GlobalConfiguration
 import liquibase.LabelExpression
 import liquibase.Liquibase
-import liquibase.Scope
-import liquibase.UpdateSummaryEnum
-import liquibase.UpdateSummaryOutputEnum
 import liquibase.command.CommandScope
 import liquibase.command.core.StatusCommandStep
-import liquibase.command.core.TagCommandStep
-import liquibase.command.core.UpdateCommandStep
-import liquibase.command.core.UpdateSqlCommandStep
-import liquibase.command.core.helpers.ChangeExecListenerCommandStep
-import liquibase.command.core.helpers.DatabaseChangelogCommandStep
-import liquibase.command.core.helpers.DbUrlConnectionArgumentsCommandStep
-import liquibase.command.core.helpers.ShowSummaryArgument
 import liquibase.database.Database
 import liquibase.database.DatabaseFactory
 import liquibase.database.OfflineConnection
 import liquibase.database.jvm.JdbcConnection
-import liquibase.io.WriterOutputStream
 import liquibase.resource.ResourceAccessor
 import net.corda.db.admin.DbChange
 import net.corda.db.admin.LiquibaseSchemaMigrator
@@ -52,10 +40,7 @@ class LiquibaseSchemaMigratorImpl(
                 .getInstance()
                 .findCorrectDatabaseImplementation(OfflineConnection(url, resourceAccessor))
         },
-    private val commandScopeFactory: (commandNames: Array<String>) -> CommandScope = { commandNames ->
-        @Suppress("SpreadOperator")
-        CommandScope(*commandNames)
-    }
+    private val liquibaseManager: LiquibaseManager = LiquibaseManager()
 ) : LiquibaseSchemaMigrator {
     companion object {
         // default schema
@@ -148,26 +133,7 @@ class LiquibaseSchemaMigratorImpl(
                         "DB Schema for ${database.connection.catalog}"
             )
 
-            val scopeObjects = mapOf(
-                Scope.Attr.resourceAccessor.name to lb.resourceAccessor
-            )
-            Scope.child(scopeObjects) {
-                if (sql != null) {
-                    commandScopeFactory(UpdateSqlCommandStep.COMMAND_NAME).configure(lb, tag).also {
-                        it.setOutput(
-                            WriterOutputStream(
-                                sql,
-                                GlobalConfiguration.OUTPUT_FILE_ENCODING.currentValue
-                            )
-                        )
-                        it.execute()
-                    }
-                } else {
-                    commandScopeFactory(UpdateCommandStep.COMMAND_NAME).configure(lb, tag).also {
-                        it.execute()
-                    }
-                }
-            }
+            liquibaseManager.update(lb, sql, tag)
             log.info("${database.connection.catalog} DB schema update complete")
         }
     }
@@ -189,20 +155,7 @@ class LiquibaseSchemaMigratorImpl(
             )
 
             log.info("Retrieving ${database.databaseProductName} DB Schema")
-            val scopeObjects = mapOf(
-                Scope.Attr.resourceAccessor.name to lb.resourceAccessor
-            )
-            Scope.child(scopeObjects) {
-                commandScopeFactory(UpdateSqlCommandStep.COMMAND_NAME).configure(lb, null).also {
-                    it.setOutput(
-                        WriterOutputStream(
-                            sql,
-                            GlobalConfiguration.OUTPUT_FILE_ENCODING.currentValue
-                        )
-                    )
-                    it.execute()
-                }
-            }
+            liquibaseManager.update(lb, sql)
             File(offlineChangeLogFileName).delete()
         }
     }
@@ -234,23 +187,4 @@ class LiquibaseSchemaMigratorImpl(
             log.info("${database.connection.catalog} DB schema rollback complete")
         }
     }
-}
-
-private fun CommandScope.configure(lb: Liquibase, tag: String?): CommandScope {
-    return this.addArgumentValue(DbUrlConnectionArgumentsCommandStep.DATABASE_ARG, lb.database)
-        .addArgumentValue(UpdateCommandStep.CHANGELOG_ARG, lb.databaseChangeLog)
-        .addArgumentValue(UpdateCommandStep.CHANGELOG_FILE_ARG, lb.changeLogFile)
-        .addArgumentValue(UpdateCommandStep.CONTEXTS_ARG, Contexts().toString())
-        .addArgumentValue(UpdateCommandStep.LABEL_FILTER_ARG, LabelExpression().originalString)
-        .addArgumentValue(
-            ChangeExecListenerCommandStep.CHANGE_EXEC_LISTENER_ARG,
-            lb.defaultChangeExecListener
-        )
-        .addArgumentValue(ShowSummaryArgument.SHOW_SUMMARY_OUTPUT, UpdateSummaryOutputEnum.LOG)
-        .addArgumentValue(
-            DatabaseChangelogCommandStep.CHANGELOG_PARAMETERS,
-            lb.changeLogParameters
-        )
-        .addArgumentValue(ShowSummaryArgument.SHOW_SUMMARY, UpdateSummaryEnum.SUMMARY)
-        .addArgumentValue(TagCommandStep.TAG_ARG, tag)
 }
