+++ tools/plugins/db-config/src/test/kotlin/net/corda/cli/plugins/dbconfig/SpecTest.kt
@@ -4,6 +4,7 @@ import liquibase.Liquibase
 import liquibase.command.CommandArgumentDefinition
 import liquibase.command.CommandScope
 import liquibase.database.Database
+import net.corda.db.admin.impl.LiquibaseManager
 import org.junit.jupiter.api.Test
 import org.mockito.ArgumentMatchers
 import org.mockito.kotlin.any
@@ -45,6 +46,7 @@ class SpecTest {
     private val commandScopeFactory = mock<(commandNames: Array<String>) -> CommandScope> {
         on { invoke(any()) } doReturn (commandScope)
     }
+    private val liquibaseManager = LiquibaseManager(commandScopeFactory)
 
     private val specConfig = Spec.SpecConfig(
         writerFactory = mockWriterFactory,
@@ -67,7 +69,7 @@ class SpecTest {
 
     @Test
     fun `Verify we run offline update and write the result to disk where no filter is specified`() {
-        val spec = Spec(specConfig, commandScopeFactory)
+        val spec = Spec(specConfig, liquibaseManager)
 
         spec.run()
 
@@ -80,7 +82,7 @@ class SpecTest {
 
     @Test
     fun `Verify we run offline update and write the result to disk only once with a filter`() {
-        val spec = Spec(specConfig, commandScopeFactory)
+        val spec = Spec(specConfig, liquibaseManager)
 
         spec.schemasToGenerate = listOf("messagebus")
 
@@ -95,7 +97,7 @@ class SpecTest {
 
     @Test
     fun `Verify we delete the changelog file if clear is specified`() {
-        val spec = Spec(specConfig, commandScopeFactory)
+        val spec = Spec(specConfig, liquibaseManager)
 
         spec.clearChangeLog = true
 
@@ -106,7 +108,7 @@ class SpecTest {
 
     @Test
     fun `Verify we delete the changelog file at a custom location if clear is specified`() {
-        val spec = Spec(specConfig, commandScopeFactory)
+        val spec = Spec(specConfig, liquibaseManager)
 
         spec.clearChangeLog = true
         spec.databaseChangeLogFile = Path.of(CUSTOM_PATH)
@@ -118,7 +120,7 @@ class SpecTest {
 
     @Test
     fun `Verify specifying jdbc url attempts to connect to a live database`() {
-        val spec = Spec(specConfig, commandScopeFactory)
+        val spec = Spec(specConfig, liquibaseManager)
 
         spec.jdbcUrl = JDBC_URL
         spec.user = USER
@@ -136,7 +138,7 @@ class SpecTest {
 
     @Test
     fun `Verify specifying statemanager schema will generate only statemanager sql`() {
-        val spec = Spec(specConfig, commandScopeFactory)
+        val spec = Spec(specConfig, liquibaseManager)
 
         spec.jdbcUrl = JDBC_URL
         spec.user = USER
